updated parameter selection
